got rid of the tank.util.authentication module and moved is_script_user_authenticated to the authentication module
